Fix generation bugs

+++ include/Generator/GenChamber.hpp
@@ -15,7 +15,7 @@ public:
 
     GenChamber(
         // The area limiter which tells generator whether this cooridinate could be use
-        std::function<bool(glm::vec2)> limiter,
+        std::function<bool(glm::ivec2)> limiter,
 
         std::shared_ptr<MapBlueprint> blueprint,
         std::shared_ptr<RandomChoose> random
@@ -29,12 +29,12 @@ public:
     ~GenChamber() = default;
     
     virtual void Generate() = 0;
-    virtual std::vector<glm::vec2> GetAvailableCooridinate() = 0;
+    virtual std::vector<glm::ivec2> GetAvailableCooridinate() = 0;
     
 protected:
-    bool isCooridinateInBound(glm::vec2 position);
+    bool isCooridinateInBound(glm::ivec2 position);
 
-    std::function<bool(glm::vec2)> m_Limiter;
+    std::function<bool(glm::ivec2)> m_Limiter;
 
     std::shared_ptr<MapBlueprint> m_Blueprint;
     std::shared_ptr<RandomChoose> m_RandomChoose;
